requirements.txt
joblib==1.2.0
matplotlib==3.5.3
numpy==1.21.6
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.5.4.60
pandas==1.3.5
Pillow==9.4.0
Pillow==9.5.0
scikit_learn==1.0.2
streamlit==1.15.1
streamlit_option_menu==0.3.2
tensorflow==2.10.0
tensorflow_intel==2.11.0
torch==1.13.1
